refactor(web): move the bundle confirm page onto the naming and layout rules
`clients/web/docs/STYLE_GUIDE.md` requires kebab-case filenames, with
generated files as the only exception, and `docs/CONVENTIONS.md` says a
directory holding exactly one file should be flattened into the correct
parent. `src/pages/BundleConfirmPage.tsx` was the only file in the tree
breaking either rule, and it broke both: it was the sole occupant of
`src/pages/`, a directory that appears in neither doc's layout.

It now sits at `src/components/bundle-confirm-page.tsx`, beside
`dictation-overlay-page.tsx`. That is the existing home for a full page
that belongs to no domain and reads its state shape from
`runtime/is-electron`, which is what both of these Electron shell screens
are.

The exported component name is unchanged, so the only call site is the
lazy import in `routes.tsx`. The route still resolves to its own chunk
(`bundle-confirm-page-*.js` in a production build), so the code splitting
is intact.
